[otp-field] Remove Group part
