+++ docs/src/app/(docs)/react/components/otp-field/demos/alphanumeric/css-modules/index.tsx
@@ -9,17 +9,17 @@ export default function OTPFieldAlphanumericDemo() {
   const descriptionId = `${id}-description`;
 
   return (
-    <OTPField.Root
-      id={id}
-      length={CODE_LENGTH}
-      validationType="alphanumeric"
-      aria-describedby={descriptionId}
-      className={styles.Field}
-    >
+    <div className={styles.Field}>
       <label htmlFor={id} className={styles.Label}>
         Recovery code
       </label>
-      <OTPField.Group className={styles.Group}>
+      <OTPField.Root
+        id={id}
+        length={CODE_LENGTH}
+        validationType="alphanumeric"
+        aria-describedby={descriptionId}
+        className={styles.Root}
+      >
         {Array.from({ length: CODE_LENGTH }, (_, index) => (
           <OTPField.Input
             key={index}
@@ -27,11 +27,11 @@ export default function OTPFieldAlphanumericDemo() {
             aria-label={`Character ${index + 1} of ${CODE_LENGTH}`}
           />
         ))}
-      </OTPField.Group>
+      </OTPField.Root>
       <p id={descriptionId} className={styles.Description}>
         Accept letters and numbers for backup codes such as{' '}
         <span className={styles.Code}>A7C9XZ</span>.
       </p>
-    </OTPField.Root>
+    </div>
   );
 }
