+++ docs/src/app/(docs)/react/components/otp-field/demos/hero/css-modules/index.module.css
@@ -14,7 +14,7 @@
   color: var(--color-gray-900);
 }
 
-.Group {
+.Root {
   display: flex;
   gap: 0.5rem;
   width: 100%;
