+++ docs/src/app/(docs)/react/components/page.mdx
@@ -1274,7 +1274,6 @@ A one-time password input composed of individual character slots.
     - Custom sanitization
   - API reference
     - Root
-    - Group
     - Input
 - Exports:
   - OTP Field - Root
@@ -1283,10 +1282,7 @@ A one-time password input composed of individual character slots.
   - OTP Field - Input
     - Props: className, render, style
     - Data Attributes: data-complete, data-dirty, data-disabled, data-filled, data-focused, data-invalid, data-readonly, data-required, data-touched, data-valid
-  - OTP Field - Group
-    - Props: className, render, style
-    - Data Attributes: data-complete, data-dirty, data-disabled, data-filled, data-focused, data-invalid, data-readonly, data-required, data-touched, data-valid
-- Types: OTPField.Group.Props, OTPField.Group.State, OTPField.Input.Props, OTPField.Input.State, OTPField.Root.ChangeEventDetails, OTPField.Root.ChangeEventReason, OTPField.Root.CompleteEventDetails, OTPField.Root.CompleteEventReason, OTPField.Root.InvalidEventDetails, OTPField.Root.InvalidEventReason, OTPField.Root.Props, OTPField.Root.State, OTPField.Root.ValidationType
+- Types: OTPField.Input.Props, OTPField.Input.State, OTPField.Root.ChangeEventDetails, OTPField.Root.ChangeEventReason, OTPField.Root.CompleteEventDetails, OTPField.Root.CompleteEventReason, OTPField.Root.InvalidEventDetails, OTPField.Root.InvalidEventReason, OTPField.Root.Props, OTPField.Root.State, OTPField.Root.ValidationType
 
 </details>
