+++ packages/react/src/otp-field/index.parts.ts
@@ -1,3 +1,2 @@
 export { OTPFieldRoot as Root } from './root/OTPFieldRoot';
-export { OTPFieldGroup as Group } from './group/OTPFieldGroup';
 export { OTPFieldInput as Input } from './input/OTPFieldInput';
