+++ packages/react/src/otp-field/index.ts
@@ -1,5 +1,4 @@
 export * as OTPField from './index.parts';
 
 export type * from './root/OTPFieldRoot';
-export type * from './group/OTPFieldGroup';
 export type * from './input/OTPFieldInput';
